% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_collect2.R
\name{bq_cloud_table_import}
\alias{bq_cloud_table_import}
\title{BigQuery Cloud Table Import}
\usage{
bq_cloud_table_import(bq_temp, gcs_mount, cleanup = F)
}
\arguments{
\item{bq_temp}{A computed temporary (or otherwise) BigQuery Table}

\item{gcs_mount}{A gcsfuse mount on the destination}

\item{cleanup}{Should exported table be removed from cloud bucket automatically: TRUE/FALSE}
}
\value{
A dataframe
}
\description{
BigQuery Cloud Table Import
}
\keyword{internal}
