% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_bq_data.R
\docType{data}
\name{connect_template}
\alias{connect_template}
\title{Connect Template}
\format{
A character vector with 16 elements
}
\usage{
connect_template
}
\description{
A character vector containing a connection function template
}
\keyword{datasets}
