% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_collect.R
\name{eureka_collect}
\alias{eureka_collect}
\title{eureka_collect}
\usage{
eureka_collect(.data)
}
\arguments{
\item{.data}{A tbl containing connection information}
}
\value{
A local data frame
}
\description{
This version of collect will actually honor the bigint variable passed by a DBI connection object
}
