% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_collect2.R
\name{eureka_collect2}
\alias{eureka_collect2}
\title{Eureka Collect 2}
\usage{
eureka_collect2(.data, use_bucket = 'no', gcs_mount, cleanup = F)
}
\arguments{
\item{.data}{A tbl containing connection information}

\item{use_bucket}{Yes/No If you know the result is too large, select 'Yes'}

\item{gcs_mount}{A gcsfuse mount on the destination}

\item{cleanup}{Should exported table be removed from cloud bucket automatically: TRUE/FALSE}
}
\value{
A dataframe
}
\description{
Eureka Collect 2
}
