% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_dbConnect.R
\name{eureka_dbConnect}
\alias{eureka_dbConnect}
\title{eureka_dbConnect}
\usage{
eureka_dbConnect(
  project_id,
  dataset_id = NULL,
  bigint_type = "character",
  google_account_type
)
}
\arguments{
\item{project_id}{A Google Cloud Project ID}

\item{dataset_id}{A Google BigQuery Dataset ID}

\item{bigint_type}{CAST large integers to this type \itemize{
\item integer
\item integer64
\item numeric
\item character (usually the best choice)
}}

\item{google_account_type}{Google account type or GSuite domain}
}
\value{
A DBI connection object
}
\description{
This function will assist in creating a DBI connection object by allowing
you to specify which cached Google credentials to use.
}
\examples{
# Create a DBI connection object with user credentials tagged with 'hdcuser'
\dontrun{
eureka_dbConnect(project_id = "sandbox-nlp",
  dataset_id = "MIMIC3",
  bigint_type = "character",
  google_account_type = "hdcuser")
  }
}
