% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_dbCreateToken.R
\name{eureka_dbCreateToken}
\alias{eureka_dbCreateToken}
\title{eureka_dbCreateToken}
\usage{
eureka_dbCreateToken(google_account_type)
}
\arguments{
\item{google_account_type}{Google account type or GSuite domain.}
}
\value{
Saves an authenticated Google OAuth token to current user's home directory
}
\description{
This function authenticates with Google using a Wiley Lab OAuth Consent Screen. Credentials are
cached in a user's home directory and labeled according to the currently connected
linux user account and the specified Google Account Type.
}
\examples{
\dontrun{
eureka_dbCreateToken(google_account_type = "hdcuser")
}
}
