% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_inernet.R
\docType{data}
\name{eureka_internet_db}
\alias{eureka_internet_db}
\title{Eureka Internet Database}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{service}{The name of the HDC approved web-service.}
\item{cmd}{The Eureka system command used to initiate the request to open the Eureka firewall to the desired web-service.}
\item{url}{The URL corresponding to the HDC approved web-service.}
...
}
}
\usage{
eureka_internet_db
}
\description{
A dataset containing a directory of Health Data Compass approved web-services,
their corresponding Eureka firewall access commands, and the URL where the
service may be accessed.

Currently implemented services include:
\itemize{
\item{"Bioconductor"}
\item{"CRAN"}
\item{"GitHub"}
\item{"REDCap"}
}
}
\keyword{datasets}
