% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_inernet.R
\name{eureka_internet_request}
\alias{eureka_internet_request}
\title{Eureka Internet Request}
\usage{
eureka_internet_request(service)
}
\arguments{
\item{service}{A case-insensitive character vector, containing the name of the requested
Health Data Compass approved web-service.
Currently implemented services include:
\itemize{
\item{"Bioconductor"}
\item{"CRAN"}
\item{"GitHub"}
\item{"RedCAP"}
}}
}
\value{
Invisibly return a \code{tibble} containing the Health Data Compass generated
firewall request messageId, the requested service, and the time the request was
submitted.
}
\description{
This function will submit a request the Eureka firewall to allow access to select
Health Data Compass (HDC) approved web-services. It leverages the HDC provided
\verb{eureka-internet-*} flavor of system commands to submit requests only after and
verifying that a live human being is submitting the request.
}
\examples{
\dontrun{
eureka_internet_open('github')
}
}
