% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureka_inernet.R
\name{eureka_internet_status}
\alias{eureka_internet_status}
\title{Eureka Internet Status}
\usage{
eureka_internet_status(service)
}
\arguments{
\item{service}{A case-insensitive character vector, containing the name of the requested
Health Data Compass approved web-service.
Currently implemented services include:
\itemize{
\item{"Bioconductor"}
\item{"CRAN"}
\item{"GitHub"}
\item{"REDCap"}
}}
}
\value{
Invisibly return a list containing the url of the tested service, the returned HTTP
status code, the time the status was checked, and additional information if url is
successfully queried.
}
\description{
This function will check the Eureka firewall status (open/closed) to Health
Data Compass approved web-services from within R.
}
\examples{
\dontrun{
eureka_internet_status('github')
}
}
