% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_bq_data.R
\docType{data}
\name{item_template}
\alias{item_template}
\title{Item Template}
\format{
A character vector with 1 element
}
\usage{
item_template
}
\description{
A character vector containing a description ITEM template
}
\keyword{datasets}
