% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_bq_data.R
\docType{data}
\name{license}
\alias{license}
\title{License Template}
\format{
A character vector with 2 elements
}
\usage{
license
}
\description{
A character vector containing a LICENSE file template
}
\keyword{datasets}
