% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_bq_data.R
\docType{data}
\name{license_md}
\alias{license_md}
\title{License Md Template}
\format{
A character vector with 21 elements
}
\usage{
license_md
}
\description{
A character vector containing a LICENSE.md template
}
\keyword{datasets}
