% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_bq_data.R
\name{package_bq_data}
\alias{package_bq_data}
\title{Package BigQuery Data}
\usage{
package_bq_data(con, google_account_type, path)
}
\arguments{
\item{con}{An EuReka::eureka_dbConnect object}

\item{google_account_type}{Google account type or GSuite domain}

\item{path}{Where to create the data package}
}
\value{

}
\description{
Package BigQuery Data
}
