% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_bq_data.R
\docType{data}
\name{table_template}
\alias{table_template}
\title{Table Function Template}
\format{
A character vector with 19 elements
}
\usage{
table_template
}
\description{
A character vector containing a table function template
}
\keyword{datasets}
